Refactor notifications API into dedicated handlers
