+++ src/services/notifications.ts
@@ -17,26 +17,26 @@ type UpdateConsentPayload = {
 
 export const notificationService = {
   send: (data: { userId: string; type: string; title: string; message: string; data?: Record<string, unknown> }) =>
-    apiClient.request('/api/notifications?action=send', {
+    apiClient.request('/api/notifications/send', {
       method: 'POST',
       body: JSON.stringify(data)
     }),
   applicationSubmitted: (data: { applicationId: string; userId: string }) =>
-    apiClient.request('/api/notifications?action=application-submitted', {
+    apiClient.request('/api/notifications/application-submitted', {
       method: 'POST',
       body: JSON.stringify(data)
     }),
   dispatchChannel: (payload: DispatchChannelPayload) =>
-    apiClient.request('/api/notifications?action=dispatch-channel', {
+    apiClient.request('/api/notifications/dispatch-channel', {
       method: 'POST',
       body: JSON.stringify(payload)
     }),
   getPreferences: () =>
-    apiClient.request('/api/notifications?action=preferences', {
+    apiClient.request('/api/notifications/preferences', {
       method: 'GET'
     }),
   updateConsent: (payload: UpdateConsentPayload) =>
-    apiClient.request('/api/notifications?action=update-consent', {
+    apiClient.request('/api/notifications/update-consent', {
       method: 'POST',
       body: JSON.stringify(payload)
     })
